+++ app.js
@@ -11,9 +11,9 @@ let dockerProxyRouter = require('./routes/docker');
 let auth = require('./routes/auth');
 let stacksApi = require('./routes/stacksApi');
 
-let jwtMiddleware = require('./jwt');
+let jwtMiddleware = require('./lib/jwt');
 
-const db = require('./db');
+const db = require('./lib/db');
 
 const UserManager = require('./lib/user_manager');
 
@@ -21,7 +21,6 @@ let app = express();
 
 let userManager = new UserManager(db);
 initDatabase(); // block
-userManager.createUser('Fred', 'pop');
 
 // view engine setup
 app.set('views', path.join(__dirname, 'views'));
@@ -33,8 +32,6 @@ if (JWT_SECRET === undefined || JWT_SECRET === '') {
     console.warn('Warning: JWT secret not set! Change JWT_SECRET to the required JWT secret value.');
 }
 
-// uncomment after placing your favicon in /public
-// app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
 
 // CORS for all routes
 app.use(cors());
@@ -73,8 +70,12 @@ app.use(function(err, req, res, next) {
     res.render('error');
 });
 
+/**
+ * Perform database initialization operations
+ *
+ */
 async function initDatabase() {
-    console.log("Initializing database...");
+    console.log('Initializing database...');
     await userManager.initDatabase();
 }