+++ routes/auth.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = new express.Router();
 const jwt = require('jsonwebtoken');
 const UserManager = require('../lib/user_manager');
-const db = require('../db');
+const db = require('../lib/db');
 
 const userManager = new UserManager(db);