+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `/user` endpoint for CRUD operations on users ([#28])
 
 ## [0.1.0]
 ### Added
@@ -28,6 +30,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [0.0.2]: https://github.com/TripleParity/docks-api/compare/0.0.1...0.0.2
 
 [#41]: https://github.com/TripleParity/docks-api/issues/41
+[#28]: https://github.com/TripleParity/docks-api/issues/28
 [#29]: https://github.com/TripleParity/docks-api/issues/29
 
 [frisby]: https://www.frisbyjs.com/